feat: implement URL shortening service API

+++ src/main/java/com/project/url_shortening/domain/model/UrlShortener.java
@@ -29,5 +29,5 @@ public class UrlShortener {
     LocalDateTime updatedAt;
 
     @Column(nullable = false)
-    Long accessCount;
+    Integer accessCount;
 }
